+++ lib/CodeTransform/Tests/Adapter/WorseReflection/Transformer/UpdateDocblockTransformerTest.php
@@ -4,10 +4,12 @@ namespace Phpactor\CodeTransform\Tests\Adapter\WorseReflection\Transformer;
 
 use Generator;
 use Phpactor\CodeBuilder\Util\TextFormat;
+use Phpactor\CodeTransform\Adapter\DocblockParser\ParserDocblockUpdater;
 use Phpactor\CodeTransform\Adapter\WorseReflection\Transformer\UpdateDocblockTransformer;
 use Phpactor\CodeTransform\Domain\Diagnostic;
 use Phpactor\CodeTransform\Domain\SourceCode;
 use Phpactor\CodeTransform\Tests\Adapter\WorseReflection\WorseTestCase;
+use Phpactor\DocblockParser\DocblockParser;
 use Phpactor\WorseReflection\Reflector;
 
 class UpdateDocblockTransformerTest extends WorseTestCase
@@ -518,7 +520,9 @@ class UpdateDocblockTransformerTest extends WorseTestCase
                     /**
                      * @return Baz[]
                      */
-                    public function baz(): array {}
+                    public function baz(): array {
+                        return [];
+                    }
                 }
 
                 class Foobar extends Foobag {
@@ -536,7 +540,9 @@ class UpdateDocblockTransformerTest extends WorseTestCase
                     /**
                      * @return Baz[]
                      */
-                    public function baz(): array {}
+                    public function baz(): array {
+                        return [];
+                    }
                 }
 
                 class Foobar extends Foobag {
@@ -618,6 +624,96 @@ class UpdateDocblockTransformerTest extends WorseTestCase
                 trait Foobar {
 
                     /**
+                     * @return Generator<array{string,Closure(Bar): string}>
+                     */
+                    public function baz(): array
+                    {
+                        yield [
+                            'foobar',
+                            function (Bar $b): string {
+                            }
+                        ];
+                    }
+                }
+                EOT
+        ];
+
+        yield 'updates existing docblock' => [
+            <<<'EOT'
+                <?php
+
+                namespace Foo;
+
+                trait Foobar {
+                    /**
+                     *
+                     */
+                    public function baz(): array
+                    {
+                        yield [
+                            'foobar',
+                            function (Bar $b): string {
+                            }
+                        ];
+                    }
+                }
+                EOT
+            ,
+            <<<'EOT'
+                <?php
+
+                namespace Foo;
+
+                use Generator;
+
+                trait Foobar {
+                    /**
+                     *
+                     * @return Generator<array{string,Closure(Bar): string}>
+                     */
+                    public function baz(): array
+                    {
+                        yield [
+                            'foobar',
+                            function (Bar $b): string {
+                            }
+                        ];
+                    }
+                }
+                EOT
+        ];
+
+        yield 'updates existing docblock with other tags' => [
+            <<<'EOT'
+                <?php
+
+                namespace Foo;
+
+                trait Foobar {
+                    /**
+                     * @author Daniel Leech
+                     */
+                    public function baz(): array
+                    {
+                        yield [
+                            'foobar',
+                            function (Bar $b): string {
+                            }
+                        ];
+                    }
+                }
+                EOT
+            ,
+            <<<'EOT'
+                <?php
+
+                namespace Foo;
+
+                use Generator;
+
+                trait Foobar {
+                    /**
+                     * @author Daniel Leech
                      * @return Generator<array{string,Closure(Bar): string}>
                      */
                     public function baz(): array
@@ -724,7 +820,8 @@ class UpdateDocblockTransformerTest extends WorseTestCase
             $reflector,
             $this->updater(),
             $this->builderFactory($reflector),
-            new TextFormat()
+            new TextFormat(),
+            new ParserDocblockUpdater(DocblockParser::create())
         );
     }
 }