+++ lib/DocblockParser/Ast/Docblock.php
@@ -56,6 +56,17 @@ class Docblock extends Node
         }
     }
 
+    public function phpDocOpen(): ?Token
+    {
+        foreach ($this->tokens() as $token) {
+            if ($token->type === Token::T_PHPDOC_OPEN) {
+                return $token;
+            }
+        }
+
+        return null;
+    }
+
     public function prose(): string
     {
         return trim(implode('', array_map(function (Element $token): string {
@@ -71,5 +82,41 @@ class Docblock extends Node
             }
             return '';
         }, iterator_to_array($this->children, false))));
+    }
+
+    public function lastMultilineContentToken(): ?Token
+    {
+        $hasLeading = false;
+        $lastToken = null;
+        foreach ($this->tokens() as $child) {
+            if ($child->type === Token::T_ASTERISK) {
+                $hasLeading = true;
+            }
+            if ($child->type === Token::T_PHPDOC_CLOSE && $hasLeading) {
+                return $lastToken;
+            }
+            $lastToken = $child;
+        }
+
+        return null;
+    }
+
+    public function indentationLevel(): int
+    {
+        $previous = null;
+        foreach ($this->children->elements as $child) {
+            if (!$child instanceof Token) {
+                continue;
+            }
+            if ($child->type === Token::T_ASTERISK) {
+                return $previous->length();
+            }
+            if ($child->type === Token::T_PHPDOC_CLOSE) {
+                return $previous->length();
+            }
+            $previous = $child;
+        }
+
+        return 0;
     }
 }