+++ lib/DocblockParser/Ast/Node.php
@@ -48,7 +48,9 @@ abstract class Node implements Element
     }
 
     /**
-     * @return Generator<Element>
+     * @template T of Element
+     * @param class-string<T> $elementFqn
+     * @return ($elementFqn is null ? Generator<Element> : Generator<T>)
      */
     public function descendantElements(?string $elementFqn = null): Generator
     {
@@ -64,6 +66,10 @@ abstract class Node implements Element
         }
     }
 
+    /**
+     * @template T of Element
+     * @param class-string<T> $elementFqn
+     */
     public function hasDescendant(string $elementFqn): bool
     {
         foreach ($this->descendantElements($elementFqn) as $element) {
@@ -140,6 +146,11 @@ abstract class Node implements Element
         return false;
     }
 
+    public function length(): int
+    {
+        return $this->end() - $this->start();
+    }
+
     /**
      * @param iterable<Element|array<Element>> $nodes
      *
@@ -188,11 +199,6 @@ abstract class Node implements Element
         }
 
         return 0;
-    }
-
-    private function length(): int
-    {
-        return $this->end() - $this->start();
     }
 
     /**